Move FAQ page logic to resources module
