+++ data/faq.json
@@ -102,7 +102,7 @@
             "questions": [
                 {
                     "question": "How should I reference the name ATT&CK?",
-                    "answer": "<p>Both MITRE ATT&CK<sup>&reg;</sup> and ATT&CK<sup>&reg;</sup> are registered trademarks of The MITRE Corporation.</p><ul><li>Your first references in writing must include \"MITRE\" preceding \"ATT&CK<sup>&reg;</sup>\" - but subsequently should just reference \"ATT&CK\" (no registered trademark symbol required).<ul><li>Example of a first reference: <i>MITRE ATT&CK<sup>&reg;</sup> is a curated knowledge base and model for cyber adversary behavior...</i></li><li>Example of subsequent reference: <i>ATT&CK is useful for understanding security risk against known adversary behavior...</i></li></ul><li>A headline should <i>always</i> reference \"MITRE ATT&CK\" together (never only \"ATT&CK<sup>&reg;</sup>\").</li><li>Always capitalize \"ATT&CK\" to distinguish it from the surrounding text.</li><li>Do not modify the trademark, such as through hyphenation or abbreviation. For example, \"ATT&CK'd!\", \"Plan-of-ATT&CK\", \"ATTK\".</li><li>You may not display the ATT&CK trademark in any manner that implies an affiliation with, sponsorship, or endorsement by MITRE, or in a manner that can be reasonably interpreted to suggest third party content represents the views and opinions of MITRE or MITRE personnel, unless those third parties receive express permission from MITRE.</li><li>You may not use ATT&CK in your product names, service names, trademarks, logos, or company names.</li></ul> For more information please visit our <a href=\"/resources/brand\"> Brand Logo Guide </a>"
+                    "answer": "<p>Both MITRE ATT&CK<sup>&reg;</sup> and ATT&CK<sup>&reg;</sup> are registered trademarks of The MITRE Corporation.</p><p>For more information please visit our <a href=\"/resources/brand\">Brand Guide</a></p>"
                 },
                 {
                     "question": "Where can I download the MITRE ATT&CK logo?",
