+++ data/resources_navigation.json
@@ -1,76 +1,91 @@
 {
-            "name": "Resources",
-            "id": "resources",
-            "children": [
-                {"name": "General Information",
-                "id": "general-info",
-                "path": "/resources/",
-                "children": []
-                },
-                {"name": "Getting Started",
-                "id": "getting-started",
-                "path": "/resources/getting-started/",
-                "children": []
-                },
-                {
-                "name": "Training",
-                "id": "training",
-                "path": "/resources/training/",
-                "children": [{"name": "CTI Training", "id": "cti", "path": "/resources/training/cti/", "children": []}]
-                },
-                {
-                    "name": "ATT&CKcon",
-                    "id": "attackcon",
-                    "path": "/resources/attackcon/",
-                    "children": []
-                },
-                {
-                "name": "Working with ATT&CK",
-                "id": "working-with-attack",
-                "path": "/resources/working-with-attack/",
-                "children": []
-                },
-                {
-                "name": "FAQ",
-                "id": "faq",
-                "path": "/resources/faq/",
-                "children": []
-                },
-                {
-                "name": "Updates",
-                "id": "updates",
-                "path": "/resources/updates/",
-                "children": []
-                },
-                {
-                "name": "Related Projects",
-                "id": "related-projects",
-                "path": "/resources/related-projects/",
-                "children": []
-                },
-                {
-                    "name": "Brand Guide",
-                    "id": "brand",
-                    "path": "/resources/brand/",
-                    "children": []
-                },
-                {
-                    "name": "Privacy Policy",
-                    "id": "privacy",
-                    "path": "/resources/privacy/",
-                    "children": []
-                },
-                {
-                    "name": "Terms of Use",
-                    "id": "terms-of-use",
-                    "path": "/resources/terms-of-use/",
-                    "children": []
-                },
-                {
-                    "name": "Changelog",
-                    "id": "changelog",
-                    "path": "/resources/changelog.html",
-                    "children": []
-                }
-            ]
+  "name": "Resources",
+  "id": "resources",
+  "children": [
+    {
+      "name": "General Information",
+      "id": "general-info",
+      "path": "/resources/",
+      "children": []
+    },
+    {
+      "name": "Getting Started",
+      "id": "getting-started",
+      "path": "/resources/getting-started/",
+      "children": []
+    },
+    {
+      "name": "Training",
+      "id": "training",
+      "path": "/resources/training/",
+      "children": [
+        {
+          "name": "CTI Training",
+          "id": "cti",
+          "path": "/resources/training/cti/",
+          "children": []
+        }
+      ]
+    },
+    {
+      "name": "ATT&CKcon",
+      "id": "attackcon",
+      "path": "/resources/attackcon/",
+      "children": []
+    },
+    {
+      "name": "Working with ATT&CK",
+      "id": "working-with-attack",
+      "path": "/resources/working-with-attack/",
+      "children": []
+    },
+    {
+      "name": "FAQ",
+      "id": "faq",
+      "path": "/resources/faq/",
+      "children": []
+    },
+    {
+      "name": "Updates",
+      "id": "updates",
+      "path": "/resources/updates/",
+      "children": []
+    },
+    {
+      "name": "Versions of ATT&CK",
+      "id": "versions",
+      "path": "/resources/versions/",
+      "children": []
+    },
+    {
+      "name": "Related Projects",
+      "id": "related-projects",
+      "path": "/resources/related-projects/",
+      "children": []
+    },
+    {
+      "name": "Brand Guide",
+      "id": "brand",
+      "path": "/resources/brand/",
+      "children": []
+    },
+    {
+      "name": "Privacy Policy",
+      "id": "privacy",
+      "path": "/resources/privacy/",
+      "children": []
+    },
+    {
+      "name": "Terms of Use",
+      "id": "terms-of-use",
+      "path": "/resources/terms-of-use/",
+      "children": []
+    },
+    {
+      "name": "Changelog",
+      "id": "changelog",
+      "path": "/resources/changelog.html",
+      "children": []
+    }
+  ]
 }
