+++ modules/website_build/website_build_config.py
@@ -66,11 +66,5 @@ js_build_uuid = Template('var build_uuid = "${build_uuid}";\n')
 # Path for templates
 website_build_templates_path = "modules/website_build/templates/"
 
-# FAQ md
-faq_md = (
-    "Title: Frequently Asked Questions\n" "Template: general/faq-overview\n" "save_as: resources/faq/index.html\n" "data: "
-)
-
 # CHANGELOG md
 changelog_md = "Title: Changelog\n" "Template: website_build/changelog\n" "save_as: resources/changelog.html\n\n"
-
